fix incorrect bounds when generating random s
The random, per-message value s, needs to be 1<s<q.

Without this fix, the code can potentially pick 1 or q. Picking q is
harmless (s2 will be 0 and the code loops with a fresh value). Picking 1
is disastrous as it reveals the private key.

(note: rfc6979 recommends using a deterministic value for s)